+++ src/app.module.ts
@@ -8,7 +8,7 @@ import { AuthModule } from './auth/auth.module';
 import { UsersModule } from './users/users.module';
 import { RolesModule } from './roles/roles.module';
 import { ClientsModule } from './clients/clients.module';
-import { ChemicalToiletsModule } from './chemical_toilets/chemical-toilets.module';
+import { ChemicalToiletsModule } from './chemical_toilets/chemical_toilets.module';
 import { ToiletMaintenanceModule } from './toilet_maintenance/toilet_maintenance.module';
 import { VehiclesModule } from './vehicles/vehicles.module';
 import { VehicleMaintenanceModule } from './vehicle_maintenance/vehicle-maintenance.module';
