+++ src/chemical_toilets/chemical_toilets.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ChemicalToilet } from './entities/chemical-toilet.entity';
-import { ChemicalToiletsService } from './chemical-toilets.service';
-import { ChemicalToiletsController } from './chemical-toilets.controller';
+import { ChemicalToilet } from './entities/chemical_toilet.entity';
+import { ChemicalToiletsService } from './chemical_toilets.service';
+import { ChemicalToiletsController } from './chemical_toilets.controller';
 
 @Module({
   imports: [TypeOrmModule.forFeature([ChemicalToilet])],
