+++ src/chemical_toilets/chemical_toilets.service.ts
@@ -3,7 +3,7 @@ import { CreateChemicalToiletDto } from './/dto/create-chemical_toilet.dto';
 import { UpdateChemicalToiletDto } from './dto/update-chemical.toilet.dto';
 import { FilterChemicalToiletDto } from './dto/filter-chemical_toilet.dto';
 import { InjectRepository } from '@nestjs/typeorm';
-import { ChemicalToilet } from './entities/chemical-toilet.entity';
+import { ChemicalToilet } from './entities/chemical_toilet.entity';
 import { Repository } from 'typeorm';
 
 @Injectable()
