+++ src/toilet_maintenance/toilet_maintenance.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { ToiletMaintenanceService } from './toilet_maintenance.service';
 import { ToiletMaintenanceController } from './toilet_maintenance.controller';
 import { ToiletMaintenance } from './entities/toilet_maintenance.entity';
-import { ChemicalToilet } from '../chemical_toilets/entities/chemical-toilet.entity';
+import { ChemicalToilet } from '../chemical_toilets/entities/chemical_toilet.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([ToiletMaintenance, ChemicalToilet])],
